fix :beetle: with prop and propPath and remove safe from core
