+++ src/Maybe/prop.spec.js
@@ -42,5 +42,11 @@ test('prop function', t => {
   t.equals(arrBad.option('nothing'), 'nothing', 'returns a Nothing when index is not found')
   t.equals(arrNull(arr).option('nothing'), null, 'returns a Just null when index is found and value is null')
 
+  const fn =
+    x => prop('key', x).option('nothing')
+
+  t.equals(fn(undefined), 'nothing', 'returns Nothing when data is undefined')
+  t.equals(fn(null), 'nothing', 'returns Nothing when data is null')
+
   t.end()
 })
