+++ src/Maybe/propPath.js
@@ -2,15 +2,17 @@
 /** @author Ian Hofmann-Hicks (evil) */
 
 const Maybe = require('../core/Maybe')
+const { Nothing, Just } = Maybe
+
 const curry = require('../core/curry')
+const isArray = require('../core/isArray')
 const isDefined = require('../core/isDefined')
 const isInteger = require('../core/isInteger')
+const isNil= require('../core/isNil')
 const isString = require('../core/isString')
-const isArray = require('../core/isArray')
-const safe = require('../core/safe')
 
-const lift =
-  safe(isDefined)
+const lift = x =>
+  isDefined(x) ? Just(x) : Nothing()
 
 // propPath : [ String | Number ] -> a -> Maybe b
 function propPath(keys, target) {
@@ -18,6 +20,9 @@ function propPath(keys, target) {
     throw new TypeError('propPath: Array of strings or integers required for first argument')
   }
 
+  if(isNil(target)) {
+    return Nothing()
+  }
   return keys.reduce((maybe, key) => {
     if(!(isString(key) || isInteger(key))) {
       throw new TypeError('propPath: Array of strings or integers required for first argument')
