+++ src/Maybe/safe.js
@@ -1,8 +1,9 @@
 /** @license ISC License (c) copyright 2016 original and current authors */
 /** @author Ian Hofmann-Hicks (evil) */
 
-const _safe = require('../core/safe')
 const Pred = require('../core/types').proxy('Pred')
+const { Nothing, Just } = require('../core/Maybe')
+const predOrFunc = require('../core/predOrFunc')
 
 const curry = require('../core/curry')
 const isFunction = require('../core/isFunction')
@@ -14,7 +15,9 @@ function safe(pred, x) {
     throw new TypeError('safe: Pred or predicate function required for first argument')
   }
 
-  return _safe(pred)(x)
+  return predOrFunc(pred, x)
+    ? Just(x)
+    : Nothing()
 }
 
 module.exports = curry(safe)
