+++ src/Maybe/safe.spec.js
@@ -32,3 +32,31 @@ test('safe helper', t => {
 
   t.end()
 })
+
+test('safe predicate function', t => {
+  const pred = x => !!x
+
+  const f = safe(pred)
+
+  const fResult = f(false).option('nothing')
+  const tResult = f('just').option('nothing')
+
+  t.equals(fResult, 'nothing', 'returns a Nothing when false')
+  t.equals(tResult, 'just', 'returns a Just when true')
+
+  t.end()
+})
+
+test('safe Pred', t => {
+  const pred = Pred(x => !!x)
+
+  const f = safe(pred)
+
+  const fResult = f(0).option('nothing')
+  const tResult = f('just').option('nothing')
+
+  t.equals(fResult, 'nothing', 'returns a Nothing when false')
+  t.equals(tResult, 'just', 'returns a Just when true')
+
+  t.end()
+})
